Refactor metrics and analysis in citool to distinguish them better
